Fixed CMake OpenMP options
OpenMP compiler flags and definitios are only passed when ENABLE_OPENMP=ON,
and OpenMP is not available for Macs. Changed the Setup OpenMP before
include(FREUDCFlagsSetup) so ENABLE_OPENMP has a chance to be set. Tested
cmake and ccmake, as well as with no OpenMP definition and ENABLE_OPENMP=ON
and OFF. They all behave as predicted.