Update cdef class Box constructor.
This enforces the use of the new API and completely
deprecates the old API.

Based on the arguments passed to the constructor, a DeprecationWarning
is triggered when the user is providing valid arguments, but probably
used the old API without keyword arguments.
An invalid set of arguments raises a ValueError.