Cleaned up voronoi.computeNeighbors
There was some verbose code with the purpose (I think) of working
around the elements of a list of lists being aliased to each other. I
changed these to not constantly remake the lists.

There was also some questionable logic with regards to what a -1 in
scipy's ridge_points means. In particular, the code was assuming that
the buffer distance would always be large enough for all original
particles to not have 'open' voronoi cells when qhull gets the
replicated, system with open boundaries.