Fix order.LocalDescriptors behavior when given neighbor lists
The intent was always to be able to pass in a given number of maximum
neighbors to consider, regardless of how many bonds were available for
each particle. This would allow one to use
LocalDescriptors.computeNList() to compute the neighbor list once for
the maximum number of neighbors one wanted to consider and then
quickly compute the descriptors for a range of neighborhood sizes, for
example. In the midst of the neighbor list revamp, it became
impossible to pass an external neighbor list and the computation
method simply worked on all of the bonds it was given. This restores
the ability to dictate the maximum neighbor size via the num_neighbors
argument to LocalDescriptors.compute.

This change also modifies the API: to use the neighbor list generated
when computeNList() is called, pass 'precomputed' as the nlist
argument to compute()