Registration is now boost-free.
The registration code has been re-written. The new version does not rely
on boost geometry, and instead performs a similar operation of finding
nearest neighbors using only standard library containers. This should be
fundamentally the same algorithm as before, and should produce the same
results. The tests for MatchEnv still pass, and the code appears to be
slightly faster. This may or may not improve the speed of real analyses,
since the number of neighbors in the MatchEnv tests is usually 6 and
real systems may have 12 (typical) or as high as 22 neighbors in an
environment.