Added Cython wrapper for Steinhardt class.
Compiles with --ENABLE-CYTHON flag on install. Some logic behind getting
the correct order parameter had to be on the Cython side since Ql and Wl
are of different types. Therefore, I defined a Python function that
selectively calls the right Cython wrapped getter on the c++ side to be
used by the Python API.