Update msd.pyx
Not sure the exact way to re-phrase this but it is never an issue to mix GPL with another license unless you want to distribute it. The copyleft aspect is only "triggered" when distributing code. So you don't have to license your code GPL to _use_ some GPL code, you just can't later distribute it without following the GPL license. I hope that is clear, I'm not sure the changes in the doc address that perfectly, but it is not an issue to use the FFTW library in a non-GPL code. I don't want someone to think "welp I don't want to license my code GPL and I don't want to pay for the library" when they have no intention for the code to leave their laptop.