Enable scikit-build on Windows (without static members) (#676)
Some minor modifications to the CMake scripts are necessary to enable building on Windows. Additionally, now that locality (and all of the C++ code) is being built once and linked to as a single shared library by all of the Cython modules, we need to make some minor modifications to the linkage of certain objects to avoid complications on Windows. In particular, static class members (like the ITERATOR_TERMINATOR) are now namespace-level constants to avoid the need for various MSVC-specific preprocessor directives.