Modify the shear matrix used to transform the fft
This change gives me diffraction patterns that preserve the symmetries
of the crystal structure, even for different numbers of unit cells in
each direction, at least for a square lattice looking straigh down the
z-axis. I'm pretty sure the k-values are still wrong, and need to use
the same factor I used in the shear matrix somehow. I'm also not sure if
that factor I scale by is correct. It seems right that it doesn't depend
on the height of a cubic box when you're looking straight on though.