Refactor diffraction plot function
Prior to this commit, the units of the diffraction plot were in
pixels...calling something like `ax.axhline(2*np.pi)` would put a
horizontal line on the 6th or 7th row of pixels from the top. This
commit changes that behavior so that the units in the image space
correlate to units in k-space...now calling `ax.axhline(2*np.pi)` will
put a horizontal line at y = 2*pi. This enables easy drawing on the
plots, which may be useful, e.g., to make sure peaks in s(k) correspond
to the actual lattice spacing in a crystalline system.

This change does not affect using the diffraction class to find where
the bright spots are, as that is more accurately done using the
`k_vectors` attribute of the class.