Make new arrays in RDF::reset
The implementation of `reduceAndReturn` does not allow the previous
pattern of calling `prepare` in `reduce`. In the new implemtnation,
`reduceAndReturn` will return the OLD array even after the new one is
created.

The solution is to override `reset`. In subclasses of
`BondHistogramCompute`, `reset` *must* call the base class implemtnation
which resets `m_historgram`. The subclass must then make new shared
arrays that are specific to that subclass. Lastly, the arrays must also
be initialized in the constructor to allow for the user to construct the
class and use it without calling `reset` first.